+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.10.1 (2024-10-14)
 - **added:** `MutBumpVecRev::{ append, into_flattened, unchecked_push(_with), as_non_null_{ptr, slice} }`
 - **fixed:** `MutBumpVecRev::extend_from_within_clone` doing nothing for ZSTs
 - **fixed:** potential UB in `MutBumpVecRev::extend_from_slice_clone` when clone panics